+++ public/app/features/dashboard-scene/serialization/transformSceneToSaveModel.ts
@@ -10,10 +10,6 @@ import {
   SceneQueryRunner,
   SceneDataTransformer,
   SceneVariableSet,
-  QueryVariable,
-  CustomVariable,
-  DataSourceVariable,
-  ConstantVariable,
 } from '@grafana/scenes';
 import {
   AnnotationQuery,
@@ -23,7 +19,6 @@ import {
   FieldConfigSource,
   Panel,
   RowPanel,
-  VariableHide,
   VariableModel,
 } from '@grafana/schema';
 import { sortedDeepCloneWithoutNulls } from 'app/core/utils/object';
@@ -36,6 +31,8 @@ import { PanelTimeRange } from '../scene/PanelTimeRange';
 import { RowRepeaterBehavior } from '../scene/RowRepeaterBehavior';
 import { ShareQueryDataProvider } from '../scene/ShareQueryDataProvider';
 import { getPanelIdForVizPanel } from '../utils/utils';
+
+import { sceneVariablesSetToVariables } from './sceneVariablesSetToVariables';
 
 export function transformSceneToSaveModel(scene: DashboardScene): Dashboard {
   const state = scene.state;
@@ -279,101 +276,4 @@ export function dataLayersToAnnotations(layers: SceneDataLayerProvider[]) {
   }
 
   return annotations;
-}
-
-export function sceneVariablesSetToVariables(set: SceneVariableSet) {
-  const variables: VariableModel[] = [];
-  for (const variable of set.state.variables) {
-    const commonProperties = {
-      name: variable.state.name,
-      label: variable.state.label,
-      description: variable.state.description,
-      skipUrlSync: Boolean(variable.state.skipUrlSync),
-      hide: variable.state.hide || VariableHide.dontHide,
-    };
-    if (variable instanceof QueryVariable) {
-      variables.push({
-        ...commonProperties,
-        current: {
-          // TODO
-          // selected: false
-          // @ts-expect-error
-          value: variable.state.value,
-          // @ts-expect-error
-          text: variable.state.text,
-        },
-        options: [],
-        query: variable.state.query,
-        datasource: variable.state.datasource,
-        sort: variable.state.sort,
-        refresh: variable.state.refresh,
-        regex: variable.state.regex,
-        allValue: variable.state.allValue,
-        includeAll: variable.state.includeAll,
-        multi: variable.state.isMulti,
-        skipUrlSync: Boolean(variable.state.skipUrlSync),
-        hide: variable.state.hide || VariableHide.dontHide,
-      });
-    } else if (variable instanceof CustomVariable) {
-      variables.push({
-        ...commonProperties,
-        current: {
-          // TODO
-          // selected: false
-          // @ts-expect-error
-          text: variable.state.value,
-          // @ts-expect-error
-          value: variable.state.value,
-        },
-        options: variable.state.options.map((option) => ({
-          text: option.label,
-          value: option.value.toString(),
-          selected: Array.isArray(variable.state.value)
-            ? Boolean(variable.state.value.find((v) => v === option.value))
-            : variable.state.value === option.value,
-        })),
-        query: variable.state.query,
-        multi: variable.state.isMulti,
-        allValue: variable.state.allValue,
-        includeAll: variable.state.includeAll,
-      });
-    } else if (variable instanceof DataSourceVariable) {
-      variables.push({
-        ...commonProperties,
-        current: {
-          // TODO
-          // selected: false
-          // @ts-expect-error
-          value: variable.state.value,
-          // @ts-expect-error
-          text: variable.state.text,
-        },
-        options: [],
-        regex: variable.state.regex,
-        query: variable.state.pluginId,
-        multi: variable.state.isMulti,
-        allValue: variable.state.allValue,
-        includeAll: variable.state.includeAll,
-      });
-    } else if (variable instanceof ConstantVariable) {
-      variables.push({
-        ...commonProperties,
-        current: {
-          // TODO
-          // selected: false
-          // @ts-expect-error
-          value: variable.state.value,
-          // @ts-expect-error
-          text: variable.state.value,
-        },
-        // @ts-expect-error
-        query: variable.state.value,
-        hide: VariableHide.hideVariable,
-      });
-    } else {
-      throw new Error('Unsupported variable type');
-    }
-  }
-
-  return variables;
 }
